:feat: add payment, order, fix filter book by category

+++ database/migrations/2024_11_29_161247_create_oder_details_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id(); 
             $table->foreignId('book_id')->constrained(); 
             $table->integer('quantity');
-            $table->decimal('price', 8, 2);
+            $table->integer('price');
             $table->foreignId('order_id')->constrained()->onDelete('cascade'); 
             $table->timestamps();
             $table->softDeletes();
